breaking: renamed SDK to Client and change top level DashJS.SDK to SDK.
