--- src/DashJS/SDK/index.ts
@@ -1,1 +1,0 @@
-export { SDK } from './SDK';
