--- src/DashJS/index.ts
@@ -1,1 +1,0 @@
-export { DashJS } from './DashJS';
