+++ src/SDK/Client/Client.d.ts
@@ -1,7 +1,7 @@
 // @ts-ignore
 import DAPIClient from "@dashevo/dapi-client"
 
-export declare namespace SDK {
+export declare namespace Client {
     interface platformOpts {
         client: DAPIClient;
         apps: object;
