+++ src/SDK/Client/Platform/methods/contracts/get.ts
