+++ src/SDK/Client/Platform/methods/documents/broadcast.ts
