+++ src/SDK/Client/Platform/methods/documents/create.ts
