+++ src/SDK/Client/Platform/methods/identities/get.ts
