+++ src/SDK/Client/Platform/methods/identities/register.ts
