+++ src/SDK/Client/Platform/methods/names/get.ts
