+++ src/SDK/Client/Platform/methods/names/register.ts
