+++ src/SDK/SDK.spec.ts
@@ -1,0 +1,12 @@
+import { expect } from 'chai';
+import SDK from "./index";
+import 'mocha';
+
+describe('SDK', () => {
+
+  it('should provide expected class', function () {
+    expect(SDK).to.have.property('Client');
+    expect(SDK.Client.name).to.be.equal('Client')
+    expect(SDK.Client.constructor.name).to.be.equal('Function')
+  });
+});
