+++ src/SDK/index.ts
@@ -1,0 +1,1 @@
+export { SDK as default } from './SDK';
